Implemented the game history screen, component styling for the items inside, began work at game details screen

+++ app/src/main/java/com/example/gameonapp/presentation/components/FitnessChip.kt
@@ -68,7 +68,7 @@ fun FitnessChip(
 @Composable
 fun FitnessTimerChip(
     modifier: Modifier = Modifier,
-    timeValue: Long,
+    timeValue: Int,
 ) {
     val textValue = formatTime(timeValue)
     Row(
